+++ resources/js/components/voyager/inputs/RepeatInput.vue
@@ -1,10 +1,14 @@
 <template>
   <div>
-    <button class="btn btn-primary mb-4" @click.prevent="add">Add {{ question.name }}</button>
+    <button class="btn btn-gray btn-sm mb-4" @click.prevent="add">Add {{ question.name }}</button>
 
     <div class="accordion" id="accordion">
-      <div class="card" v-for="(repeat, index) in repeated" :key="repeat+index">
-        <div class="card-header flex justify-between" :id="'heading' + index">
+      <div
+        class="bg-gray-100 border rounded px-4 pt-2 pb-4"
+        v-for="(repeat, index) in repeated"
+        :key="repeat+index"
+      >
+        <div class="flex justify-between" :id="'heading' + index">
           <button
             class="btn btn-link"
             type="button"
@@ -27,8 +31,8 @@
           data-parent="#accordion"
         >
           <div class="card-body">
-            <div class="form-group" v-for="q in question.form" :key="q.id" :id="q.id">
-              <label class="control-label">
+            <div class="mb-2" v-for="q in question.form" :key="q.id" :id="q.id">
+              <label class="form-label">
                 {{ q.question }}
                 <span v-show="q.required">*</span>
               </label>
@@ -53,7 +57,7 @@ import TextInput from "../inputs/TextInput.vue";
 import SelectInput from "../inputs/SelectInput.vue";
 
 export default {
-  props: ["question"],
+  props: ["question", "value"],
   data() {
     return {
       repeated: [],
@@ -64,6 +68,10 @@ export default {
     for (let i = 0, len = this.question.form.length; i < len; i++) {
       let id = this.question.form[i].id;
       this.init[id] = "";
+    }
+
+    if (this.value.length > 0) {
+      this.repeated = this.value;
     }
 
     if (this.question.rules === "required") this.add();